Add Rays
This adds a new Sprite class, `Ray`, to model the rays that the alien can
now fire when the spacebar key is hit. Rays need to know the direction
and position of the alien (to know where to start from) and they are
killed when they go outside the screen boundary.