Create Game class with game over scene
This creates a `Game` class with a `play` method. The new class has a `stage`
and a `score`, which means it can be used to progress that game to
different stages based on the score. There is also a `game_over` method
which shows an ending screen when the game is over, instead of
immediately closing the game window.